UI 배치 변경 및 이미지 적용, Setting UI 프리펩으로 만듦
