+++ Emily/Assets/##Scripts##/WsScript/Opening.cs
@@ -7,54 +7,19 @@ namespace ws
 {
     public class Opening : MonoBehaviour
     {
-        [SerializeField] private Text volmunText;
-        [SerializeField] private Slider volumeController;
         [SerializeField] private Button startButton;
-        [SerializeField] private Button settingsButton;
         [SerializeField] private Button exitButton;
-        [SerializeField] private GameObject settingsPopupScreen;
-        public void Start()
-        {
-            SoundInitialize();
-        }
-
-        private void SoundInitialize()
-        {
-            volumeController.onValueChanged.AddListener(VolumeChanged);
-            SoundManager.instance.SetSoundValue(0.5f);
-            SoundManager.instance.PlayBGM(0);
-        }
 
         public void StartButtonDown()
         {
             SoundManager.instance.PlayEffectSound(1);
             print("Start");
         }
-        public void SettingsButtonDown()
-        {
-            SoundManager.instance.PlayEffectSound(1);
-            settingsPopupScreen.SetActive(true);
-            UpdateVolumeController();
-        }
-        private void UpdateVolumeController()
-        {
-            volumeController.value = SoundManager.instance.GetBackGroundMusicVolume() * 10;
-            ChangeVolumeText(volumeController.value);
-        }
 
         public void ExitButtonDown()
         {
             SoundManager.instance.PlayEffectSound(1);
             Application.Quit();
-        }
-        public void VolumeChanged(float value)
-        {
-            ChangeVolumeText(value);
-            SoundManager.instance.SetSoundValue(value * 0.1f);
-        }
-        private void ChangeVolumeText(float volume)
-        {
-            volmunText.text = $"{volume}";
         }
     }
 }
