+++ Emily/Assets/##Scripts##/WsScript/SettingPopup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 483f2a1348c362a41871e6bfd274e7e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
