+++ Emily/Assets/Prefabs/SettingUiCanvas.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b7d7ee79d8fea554199be91255eecf59
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
