+++ Emily/Assets/Resources/BackGround.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 987e4a975f676e74784b435c9e783a8b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
