+++ Emily/Assets/Resources/Fonts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4d3c7adfc8cead040972a7e2226d8c0e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
