+++ Emily/Assets/Resources/Fonts/NanumMyeongjoExtraBold.ttf.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: d6a3c91f2e53d84498d21228f639c157
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontNames:
+  - NanumMyeongjo
+  fallbackFontReferences:
+  - {fileID: 12800000, guid: c1d3e30f22e9e14448f66d4569726b48, type: 3}
+  - {fileID: 12800000, guid: 80a5063785b31b944b2024967add6f1a, type: 3}
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
