+++ Emily/Assets/Resources/Opening.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1af3ab539c8dd4448880648018948bbd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
